Dec. update pt. 2

+++ lclab2/src/base/scalar.h
@@ -5,7 +5,7 @@
 
 namespace LC {
 	typedef double scalar;
-	typedef long double precision_scalar;
+	typedef double precision_scalar;
 	constexpr std::size_t SIZE_OF_SCALAR = sizeof(scalar);
 	constexpr std::size_t SIZE_OF_PRECISION_SCALAR = sizeof(precision_scalar);
 }
